fix(web): force reload code editor on render
If while editing a function you go to the asset panel, then go back, the
code editor will lose syntax highliting and type checking. I tried a
bunch of things, but in the end there is something with how code editor
loads itself that I wasn't able to cleanup.

So forcing the component to load again whenever it gets rendered (it
doesn't re-render spuriously) solves the issue.

Also, we prevent the anchor click from propagating when closing a
function tab, as it was was forcing a hard reload of the page.