fix(web): prevent changeset actions in frontend while fixes are running
Creating a changeset in the middle of the fix flow causes inconsistency.
This PR prevents this by checking if fixes are running, and if they are
it

   - disables the changeset button
   - prevents dropping assets on the graph
   - disables the right click menu

There are probably other paths that could create a changeset or get into
a conflict with the fix run.

A more robust solution is needed that prevents SDF itself from
performing operations that could cause inconsistency during the fix run,
but this should make for a smoother user experience.